+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # ActFramework Change Log
 
 **1.8.2**
+* Different URL variable name caused duplicate routes not been reported #561
 * `ActUnauthorized` caused `java.lang.NegativeArraySizeException` in dev mode #563
 * Download file not working properly #562
 * Error with sending large response: UT000043: Data is already being sent. You must wait for the completion callback to be be invoked before calling send() again #560